Remove offline stuff, make it fetch the json from github
