--- package.json
@@ -1,19 +1,0 @@
-{
-  "name": "discord-flags",
-  "version": "1.0.0",
-  "description": "Documenting undocumented discord flags",
-  "main": "src/build-dist.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/lewisakura/discord-flags.git"
-  },
-  "author": "lewisakura",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/lewisakura/discord-flags/issues"
-  },
-  "homepage": "https://github.com/lewisakura/discord-flags#readme"
-}
